[final-project] change table name to singular

+++ final-project/repository/interface.go
@@ -10,7 +10,7 @@ type UserRepository interface {
 	Create(context.Context, model.User) (model.User, error)
 	FindByEmail(context.Context, string) (model.User, error)
 	Update(context.Context, model.User) (model.User, error)
-	Delete(context.Context, uint) error
+	Delete(context.Context, uint64) error
 }
 
 type PhotoRepository interface {
